+++ src/backend/typeDefs/mutations/client/update.js
@@ -4,12 +4,14 @@ const updateClientTypeDefs = gql`
   input UpdateClientInput {
     _id: ID!
     description: String!
+    icon: String
   }
 
   type UpdateClientPayload {
     _id: ID!
     name: String
     description: String
+    icon: String
   }
 `