+++ src/backend/typeDefs/queries.js
@@ -220,6 +220,7 @@ const queries = gql`
     _id: String
     name: String
     description: String
+    icon: String
   }
 
   type Team {
@@ -235,6 +236,8 @@ const queries = gql`
 
   type User {
     _id: String
+    firstName: String
+    lastName: String
     username: String
     email: String
     client: Client
@@ -677,6 +680,8 @@ const queries = gql`
 
   type PathwaysAndPersonConnection {
     _id: ID
+    createdOn: DateTime
+    updatedOn: DateTime
     personId: String!
     pathwaysId: String!
     indicationIds: [String!]!