+++ src/frontend/Orion/Organizations/Pathways/Pathways.js
@@ -3,14 +3,9 @@ import { Switch, Route, Redirect } from 'react-router-dom'
 
 import Accounts from './Accounts'
 import Influencers from './Influencers'
-import NewAccounts from './NewAccounts'
 
 const Pathways = () => (
   <Switch>
-    <Route
-      path={'/orion/organizations/pathways/newaccounts'}
-      component={NewAccounts}
-    />
     <Route
       path={'/orion/organizations/pathways/accounts'}
       component={Accounts}