+++ src/frontend/Phoenix/Home/ClientsPanel/UpdateButton.js
@@ -1,6 +1,6 @@
 import React from 'react'
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome"
-import { faEdit } from "@fortawesome/free-solid-svg-icons"
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
+import { faEdit } from '@fortawesome/free-solid-svg-icons'
 
 import { UPDATE_CLIENT } from '../../../api/mutations'
 
@@ -9,11 +9,7 @@ import { GET_CLIENTS } from '../../../api/queries'
 import TextFormButton from '../../shared/TextForm/Button'
 
 const editIcon = (
-  <FontAwesomeIcon
-    size="lg"
-    icon={faEdit}
-    style={{ color: 'white' }}
-  />
+  <FontAwesomeIcon size="md" icon={faEdit} style={{ color: 'white' }} />
 )
 
 const buttonStyle = {
@@ -23,19 +19,14 @@ const buttonStyle = {
   opacity: 0.3,
 }
 
-const UpdateButton = ({
-  client: {
-    _id,
-    description,
-  }
-}) => (
+const UpdateButton = ({ client: { _id, description, icon } }) => (
   <TextFormButton
     modalTitle="Edit Client"
     buttonLabel={editIcon}
     buttonStyle={buttonStyle}
     data={{ description }}
     mutationDoc={UPDATE_CLIENT}
-    additionalFormData={{ _id }}
+    additionalFormData={{ _id, icon }}
     refetchQueries={[{ query: GET_CLIENTS }]}
   />
 )