+++ src/frontend/api/mutations/users.js
@@ -4,6 +4,8 @@ export const SELECT_USER = gql`
   mutation SelectUser($_id: String) {
     selectUser(_id: $_id) @client {
       _id
+      firstName
+      lastName
       username
       email
       emailSubscriptions {
@@ -22,6 +24,8 @@ export const CREATE_USER = gql`
   mutation CreateUser($input: CreateUserInput!) {
     createUser(input: $input) {
       _id
+      firstName
+      lastName
       username
       email
       emailSubscriptions {
@@ -40,6 +44,8 @@ export const MANAGE_CREATED_USER = gql`
   mutation ManageCreatedUser($data: JSON) {
     manageCreatedUser(data: $data) @client {
       _id
+      firstName
+      lastName
       username
       email
       emailSubscriptions {
@@ -58,6 +64,8 @@ export const DELETE_USER = gql`
   mutation DeleteUser($input: DeleteUserInput!) {
     deleteUser(input: $input) {
       _id
+      firstName
+      lastName
       username
       email
       emailSubscriptions {
@@ -76,6 +84,8 @@ export const MANAGE_DELETED_USER = gql`
   mutation ManageDeletedUser($data: JSON) {
     manageDeletedUser(data: $data) @client {
       _id
+      firstName
+      lastName
       username
       email
       emailSubscriptions {
@@ -94,6 +104,8 @@ export const UPDATE_USER = gql`
   mutation UpdateUser($input: UpdateUserInput!) {
     updateUser(input: $input) {
       _id
+      firstName
+      lastName
       username
       email
       emailSubscriptions {
@@ -112,6 +124,8 @@ export const MANAGE_UPDATED_USER = gql`
   mutation ManageUpdatedUser($data: JSON) {
     manageUpdatedUser(data: $data) @client {
       _id
+      firstName
+      lastName
       username
       email
       emailSubscriptions {