+++ src/frontend/api/typeDefs.js
@@ -2,19 +2,20 @@ import gql from 'graphql-tag'
 
 const typeDefs = gql`
   type Query {
-    clients: [Client],
-    teams: [Team],
-    users: [User],
-    selectedClient: Client,
-    selectedTeam: Team,
-    selectedUser: User,
-    stagedSitemap: Sitemap,
+    clients: [Client]
+    teams: [Team]
+    users: [User]
+    selectedClient: Client
+    selectedTeam: Team
+    selectedUser: User
+    stagedSitemap: Sitemap
   }
 
   type Client {
     _id: String # why not ID type?
     name: String
     description: String
+    icon: String
   }
 
   type Team {
@@ -60,4 +61,4 @@ const typeDefs = gql`
   # }
 `
 
-  export default typeDefs
+export default typeDefs