+++ src/frontend/components/BusinessObjectModal/PeopleModal/widgets/relational-widgets/OrganizationConnectionsWidget/ConnectionPanel/ButtonCluster.js
@@ -1,9 +1,12 @@
-import React from 'react'
+import React, { useState } from 'react'
+import Snackbar from '@material-ui/core/Snackbar'
+import MuiAlert from '@material-ui/lab/Alert'
 import _ from 'lodash'
 import { useMutation } from '@apollo/react-hooks'
 import PropTypes from 'prop-types'
 import styled from '@emotion/styled'
 import Button from 'frontend/components/Button'
+import Spinner from 'frontend/components/Spinner'
 
 import Color from 'frontend/utils/color'
 import Spacing from 'frontend/utils/spacing'
@@ -16,6 +19,19 @@ import {
   UPSERT_PATHWAYS_AND_PERSON_CONNECTION,
   DELETE_PATHWAYS_AND_PERSON_CONNECTION,
 } from 'frontend/api/mutations'
+
+const Alert = (props) => {
+  return <MuiAlert elevation={6} variant="filled" {...props} />
+}
+
+const SpinnerWrapper = styled.div({
+  width: 43,
+  height: 30,
+  display: 'flex',
+  justifyContent: 'center',
+  alignItems: 'center',
+  margin: `0 ${Spacing.S3}`,
+})
 
 const ButtonsWrapper = styled.div({
   display: 'flex',
@@ -32,6 +48,8 @@ const ButtonCluster = ({
   changeOrganization,
   connectionsData,
 }) => {
+  const [snackbarOpen, toggleSnackbar] = useState(false)
+
   orgData = stripTypename(_.cloneDeep(orgData))
 
   const {
@@ -70,28 +88,32 @@ const ButtonCluster = ({
     endQuarter,
   }
 
-  const [upsert] = useMutation(UPSERT_PATHWAYS_AND_PERSON_CONNECTION, {
-    variables: {
-      input: dataToPersist,
-    },
-    refetchQueries: [
-      { query: GET_EVENTS },
-      { query: GET_JOIN_PATHWAYS_AND_PEOPLE },
-    ],
-    awaitRefetchQueries: true,
-    onCompleted: (res) => {
-      if (isNewOrgBeingCreated) {
-        setWhetherNewOrgBeingCreated(false)
-        const newConnectionId = Object.values(res)[0]._id
-        changeOrganization(
-          connectionsData.find(({ _id }) => _id === newConnectionId)
-        )
-      }
+  const [upsert, { loading: upsertLoading }] = useMutation(
+    UPSERT_PATHWAYS_AND_PERSON_CONNECTION,
+    {
+      variables: {
+        input: dataToPersist,
+      },
+      refetchQueries: [
+        { query: GET_EVENTS },
+        { query: GET_JOIN_PATHWAYS_AND_PEOPLE },
+      ],
+      awaitRefetchQueries: true,
+      onCompleted: (res) => {
+        if (isNewOrgBeingCreated) {
+          setWhetherNewOrgBeingCreated(false)
+          const newConnectionId = Object.values(res)[0]._id
+          changeOrganization(
+            connectionsData.find(({ _id }) => _id === newConnectionId)
+          )
+        }
 
-      setWhetherUnsavedChanges(false)
-    },
-    onError: alert,
-  })
+        toggleSnackbar(true)
+        setWhetherUnsavedChanges(false)
+      },
+      onError: alert,
+    }
+  )
 
   const [deleteConnection] = useMutation(
     DELETE_PATHWAYS_AND_PERSON_CONNECTION,
@@ -123,34 +145,57 @@ const ButtonCluster = ({
   }
 
   return (
-    <ButtonsWrapper>
-      <Button
-        color={Color.WHITE}
-        onClick={cancelHandler}
-        buttonStyle={{ color: Color.GRAY_DARK, margin: `0 ${Spacing.S3}` }}
-      >
-        Cancel
-      </Button>
-      <Button
-        type="secondary"
-        onClick={upsert}
-        color={Color.GREEN}
-        buttonStyle={{ margin: `0 ${Spacing.S3}` }}
-      >
-        Save
-      </Button>
-
-      {!isNewOrgBeingCreated && (
+    <>
+      <ButtonsWrapper>
         <Button
-          buttonStyle={{ margin: `0 ${Spacing.S3}` }}
-          onClick={deleteHandler}
-          type="secondary"
-          color={Color.RED}
-          iconName="delete"
-          iconColor1={Color.RED}
-        />
-      )}
-    </ButtonsWrapper>
+          color={Color.WHITE}
+          onClick={cancelHandler}
+          buttonStyle={{ color: Color.GRAY_DARK, margin: `0 ${Spacing.S3}` }}
+        >
+          Cancel
+        </Button>
+
+        {upsertLoading ? (
+          <SpinnerWrapper>
+            <Spinner />
+          </SpinnerWrapper>
+        ) : (
+          <Button
+            type="secondary"
+            onClick={upsert}
+            color={Color.GREEN}
+            buttonStyle={{ margin: `0 ${Spacing.S3}` }}
+          >
+            Save
+          </Button>
+        )}
+
+        {!isNewOrgBeingCreated && (
+          <Button
+            buttonStyle={{ margin: `0 ${Spacing.S3}` }}
+            onClick={deleteHandler}
+            type="secondary"
+            color={Color.RED}
+            iconName="delete"
+            iconColor1={Color.RED}
+          />
+        )}
+      </ButtonsWrapper>
+
+      <Snackbar
+        anchorOrigin={{
+          vertical: 'top',
+          horizontal: 'center',
+        }}
+        open={snackbarOpen}
+        onClose={() => toggleSnackbar(false)}
+        autoHideDuration={5000}
+      >
+        <Alert onClose={() => toggleSnackbar(false)} severity="success">
+          Connection saved!
+        </Alert>
+      </Snackbar>
+    </>
   )
 }